🚑 (Migrations) Add support for Django 3.0+ by properly checking major and minor versions
When checking for "new" versions of Django, the logic required a major version
of 2+ AND a minor version of 1+, which caused 3.0 to be missed. This resulted
in an error of `psycopg2.errors.InvalidTableDefinition: multiple primary keys
for table "django_rest_passwordreset_resetpasswordtoken" are not allowed`.

I updated the check to include all versions of Django 3 as "new".